fix(portal): refactor homepage component to use new navigation and content services
